Replace agentlock with class AgentLock and BaseMaceService::synclock.  The latter is used for synchronus functions generated for certain interface classes.
New macro: DOWNGRADE() which can be used to manually downgrade the mode to
READ_MODE (anon) or NONE_MODE (nihil).

Added code to transports and scheduler to get tickets.  I forget if I did it
for groupstream but it doesn't look like it.  Calls to multicast get ticket
numbers since the thread's ticket is already used.

Note - the compiler does not yet do auto-locking=read for const transitions.

Had to make CROR::deliver locking=write, and to have SGM::forward downgrade to
NONE_MODE.  The problems were that CROR::deliver calls checkNode (a write
transition), AND CROR is also used for Scribe Heartbeat messages, which are
also write transitions.  So, to enable parallelism, I made it go to NONE_MODE
once you determine the group identifier of a SGM message (i.e. in the forward
transition).  In a true context model, this is the point at which I exect we
would enter a context anyway.  I could have put the downgrade in GTM, but it
occurred to me that in principle, GTM could also be used for both reading and
writing (in some cases), so the downgrade took place once the final deliver
was known.

Note: currently using a sham ticket class.  Need to update to Hyo's code, and
merge.