Large commit.
Subject: Change in handling of a service's "Local Address".

Previously, there was a getLocalAddress(regid) downcall which a service was
supposed to implement.  Unfortunately, it's nature meant that you could not
call it until after a service had been initialized by maceInit, which led to
certain problems for property testing.

Additionally, the lack of knowledge of a node's local address meant that
property testing in services which did not use IPV4 addresses (such as
Scribe/SplitStream) became challenging.

Not to mention, if services neglected to implement getLocalAddress(), they
would not work properly in the simulator/model checker, because they would
return the host's IP address rather than the simulated IP address.

In the new design, there is a new function localAddress() which returns a
service's local address.  From within a service, you can get a lower service's
local address using a similar downcall_localAddress(sv).

This function generally takes no registration uid, and returns a const
reference.

By default, a service's localAddress() implmementation is to get the local
address from the lower level services.  It will ask each in turn, and if they
do not match, a runtime error will occur.  If there is only one, or if they
match, it will proceed using their address.

A service can implement the localAddress() by adding a block:

local_address {

}

This contains a function body which should return a MaceKey.  It could, for
example, call localAddress() on just one lower level service, or it could
construct a SHA address (i.e. pastry/chord/etc.).

The function can be called at any time, and returns a stored MaceKey.  The
local_address block will only be evaluated once, so a node's local address is
not allowed to change at runtime.  If this limitation becomes problematic, we
can discuss alternatives, but I am aware of no services which use this
capability.  Additionally, since it does not take a registration uid parameter,
the address cannot be specialized to the caller.  Same comment about letting me
know about design issues.

As a result of this change, a number of other things have changed:

1) The model checker now allows "promoting" an address other than a node's IPV4
address to a reference to the node.  In particular, the local address can be
used.  (It assumes this only occurs if the address family is not IPV4).

2) When checking if a node is an element of the set, instead of using it's IPV4
address, it uses its local address.  This required a change to the Pastry
properties, as they did a test on whether the node was in the leafset (sorted
on IP), and now checks against the hashleafset (sorted on SHA).

3) Model checking scripts should look for localAddress() not getLocalAddress().

4) I modified all services to be compliant to the change.

5) I added a test service to test ScribeMS (at least basically).

6) I added a new application based on appmacedon (called appmacedon_test) which
uses real networking/loopback addresses to test correct functionality of
multicasting over RandTree, Scribe, and SplitStream.

7) If you hadn't already noticed, there is now the capacity for "make test",
which can be used to validate some amount of functionality after changes, that
things aren't broken.  It should in the future be considered good practice to
do both a "make" and a "make test" before a commit.  I have a machine which is
doing a nightly test of these tests so I'll know if things break.

8) There are now 10 tests which test basic rudimentary functionality of many
services and libraries.  Feel free to add new tests.  Doing so requires running
a command which either returns 0 or non-0, and specifying that command using
the cmake command ADD_TEST in a CMakeLists.txt file.  You can see
mace/application/appmacedon_test/CMakeLists.txt or
mace/application/modelchecker/CMakeLists.txt for examples.

9) The new appmacedon_test works basically the same way as appmacedon, but to
make a functional test, it calls "fork" to create copies of itself on which to
run the protocol, then the main process waits for all others... All have to
exit successfully for the test to pass.

10) As a minor change, I have enabled node state formatted output for the model
checker.  It can be turned off by setting the PRINT_STATE_FORMATTED parameter
to 0.

11) In my prior commit - mdb will take SHA 160 addresses, and colorize them.
When formatting state, it will print only the first 6 and last 2 characters
with a .. between.

12)



------------------------------------------------------------------------
r1162 | ckillian | 2011-06-30 15:57:56 -0400 (Thu, 30 Jun 2011) | 98 lines