Redirected async call to remote node.
[WARNING:] This version generates C++ code that does async call redirection. However,
the generated C++ code WILL NOT work. But this gives a preview of what will be done eventually.

This is assuming the head node's part is working (so far, it's not the case)

Next step to work on: context-to-node mapping. Also, I will incorporate a new parser that understands context.

I need to figure out how to represent the contexts.