Bug fix.
Important notice in this update:
I added a copy constructor for all AutoType classes. This is added to solve a potential bug when AsyncDispatch enqueues a message object instance.
In the old code,  the internal members are the reference to the address in the local variable. This normally works if the variable is used immediately.
However, this would not work when the object is enqueued. The referenced address is not permanet and thus the content is unexpected.

The preferred solution is to use a copy constructor to create a copy of the original object.

Other noticeable:
A few bug fixes to make LUFactorization service works.

To ease the debugging, unit_app application can be executed directly without heartbeat application.
For example: ./unit_app -service LUFactorization -run_time 0 -initprintable job_printable.spec
(run_time = 0 means execution time is indefinite)
where job_printable.spec is a context to node mapping file in human-readable format.
For example:

head: cloud02:5001
cloud02:5001 Row0 Row0Column0
cloud03:5002 Row0Column1 Row0Column1

The "head:" specifies the head node
The next two lines explicitly maps the respective contexts to the two nodes.

-----------------------
heartbeat application:

heartbeat application forks a new process of unit_app with parameters.
For example, unit_app -service LUFfactorization -run_time 0 -initcontext contextX1F0G

where -initcontext parameter specifies the file of head and the context-node mapping in serialized MaceKey format.
This temporary file is generated by heartbeat application.