Bug fix for context declaration and begin working on context-level locking.
Fixed the problem where subcontexts within a context is rendered as a subclass definition in the context class.
The new code rendereds all subcontext as well as context class as subclass of the service class.

Beginning working on context-level locking.
Specifically, I created a ContextLock class ( lib/ContextLock.h )
This class serves the purpose similar to AgentLock, but locks at each individual context class object.
ContextLock and AgentLock should be functionally similar, but implementation-wise they are different, so
I chose to create a new class instead of modifying the existing AgentLock.

In order to use ContextLock, sub toString() in Method.pm was modified so that after AgentLock, a ContextLock
local variable is created. Later, I will remove AgentLock in the codegen.