When entering a context, async call creates a write lock on the context, and read lock on all of its parent contexts.
Incorporate more code from AgentLock to ContextLock.

I originally thought of having one mutex variable for each context, but that would not be necessary.
For now, assuming there is only one mutex, because the mutex locking takes a very short amount of time,
so it would not be a bottleneck even if only one mutex is used.