Bug fix for ContextLock. Not working code!!! Don't update to this revision!!
I adapted the code from AgentLock to ContextLock, and I also wrote several implementation notice about the difference in the implementation.
Despite recognizing the difference, I have not been able to fix the bugs. So this is a bug fix for more bugs.

All in all, I think we're making steady progress. Expecting to have a very rough finished working copy by end of week. I shall be able to put
it into test on Friday, and we'll see how the ContextLock works.

After ContextLock, the next (small) thing is to handle the problem with global context, and the single maceInit() invocation problem.