Note: 1. ContextLock implementation 2. global context hack
1. I think I got ContextLock mostly right. Read the AgentLock code several times throughly to make sure I understand how it works and how I can apply it to ContextLock.

2. Added a hack to maceInit() and maceExit()
These two methods are executed in global context and executed automatically. Special treatment is necessary to make sure these two methods are only executed on the physical node responsible for the global context. If not, it would not execute user-defined code, but only initialize/release lower services.

The context mapping file can specify a global context mapping with keyword "global" now. In the .mac file, a global context is specifies as a "[]". I.e., context name is empty.