fixed a bug with state variable, changed service/LUFactorization and change ContextLock declaration in async calls.
The bug was that a variable in the context is created, a global variable of the same name is also created.

LUFactorization was changed to use 'explicit snapshot' i.e., if a variable is used in async call, it is passed explicitly as a parameter.

ContextLock declaration: With explicit snpapshot, async calls does not need to get read lock(which is effectively the snapshot) of the parent contexts.