1. fixed context serialization bug. 2. start experimenting with failure recover.
The contexts in the services were not serialized/deserialized. This update fixed the bug.

For failure recover, I added a maceResume() for all services. Essentially it's the same as maceInit, which initializes lower level (transport, etc). However, I have not had chance to think of layered services initialization (lower layer services, except TCP/UDP transport, should also be serialized for recovery)

unit_app was changed to take snapshot by storing serialized service into a file beginning with 'ssobj_', when SIGUSR2 is received. Also, if the parameter 'resumefrom' is given, unit_app resumes the service by deserializing the state variables from the file.

Preliminary result: service can be restarted in this way. The state variables, and data in the contexts  seems to be uncorrupted. The transport/network packets are not tested yet. Will do it now.

My thought: perhaps an additional MaceKey type is not necessary. Since we are now mapping contexts to physical nodes, as long as we know this mapping relationship, calls and data would work smoothly. If nodes(as well as contexts) moves to other places because of failure, we just need to update the context-node mapping. I'll experiment with it more.