Added a "SynchronousCallWait" class.
This class is included in two files .cc and .h.
The purpose of the class is to implement RPC-like functionality where the network message does not return until
the return value is sent back. However, this class does not do sending -- it merely waits. The actual sending
of the message is done elsewhere.