First attempt to implement snapshot for contexts.
The proof-of-concept mace-fullcontext implementation did not take snapshot.

This is because the applications used in mace-fullcontext does not use read-only context. Since all context accesses are write,
snapshot is unnecessary and so therefore not implemented. On the other hand, the old code for handling snapshot taking was not
compatible with mace-fullcontext. Instead of taking snapshot of the entire state, it suffices to take the snapshot of the context.

I'm not sure how we should take the snapshot of the other co-existing services.