Fixed compilation error in Mac OS. There're three changes made:
1. lib/SampleSignificantRateControl.cc
I believe this came from the mace trunk. In Mac, pow() takes two parameters: double and int types but in Linux, pow takes (double, unsigned int).

2. lib/ThreadPool.h
This code came from mace-incontext.
definition of pthread_t is different in Mac than that in Linux. A function in this file attempted to call mace::vector< pthread_t > but pthread_t in Mac OS can not be serialized and caused error.
A correct fix is to use std::vector instead.

3. lib/ContextBaseClass.h
This code came from fullcontext branch.
The constructor of ContextBaseClass initializes a pthread_key_t variable the value PTHREAD_ONCE_INIT.
However, definition of PTHREAD_ONCE_INIT in Mac OS is not compatible with class initialization list. Therefore a hack was employed to achieve the same effect in MacOS.