More code added to support context migration.
Also, strting from this revision, the runtime no long send acknowledgement in response of the packet received. This feature was needed previously to support process migration. But it's not needed for context migration. In the new design, process migration is simply an extension of context migration, and therefore the acknowledgement is obsoleted.