Big change: The compiler supports context'ed routines.
I made a big change in the compiler. Routines can be annotated with a context. When a routine is called, a direct call message is actually created to the destination context.

The usage of routine is functionally the same as 'sync' calls. Take a look at services/Simple/Simple.mac. The old 'sync' prefix is still supported, but I plan to drop it some time later. Please update the code accordingly.

A simple example of context'ed routine syntax is as follows:

state_variables{
    context R<int row>{
        init_cells;
    }
}
routines{
    [R<row>] int getInitCells( int row ){
        return init_cells;
    }
}