Fixed a few bugs.
In several message handlers, ticket are created but not used therefore the process can not continue.

TODO:
Timer handler not done. Process crashed in timer handler because it tries to access ThreadStructure::myContext() which is NULL in timer handler. plus, timer handler is not contexted now.
In async call handler, ThreadStructure::myContext() is defined.