Adding support for context'ed timer.
In this revision, when a timer foo expires, expire_foo() is called. But instead of the original behavior which simply takes place locally, it is replaced with an async_foo() like call. This call creates messages that is sent to head to create a new event.

In the context'ed timer, the original expire_foo() is renamed to real_expire_foo()