Context Timer implementation.
In this revision, when a timer foo is fired, instead of calling  expire_foo(), it calls scheduler_foo() which sends a message to head node. The head node creates a new event and relay the message to the destination node.

In this code, it is assumed the physical node where the context of the timer is in is responsible for firing the new timer event request. Also, for simplicity, to schedule a timer, it is assumed the process has to be in the context of the timer.