Added GameOfLife into services. Fixed a few bugs in the GameOfLife services, as well as minor fix in compiler.
GameOfLife service is now compatible with the current compiler. (It compiles, but not tested yet)

Minor fix in compiler:
(1) Ignore context annotation when context of a transition is defined as either '__internal', '__anon' or '__null'
(2) Better error message when the annotated context is not declared.
(3) Declare context classes after autotype classes so that auto types can also be used inside contexts.