Clean up the filters code:
Remove GenericFilter, which has long been an artifical inheritance.  No Mace
code actually uses GenericFilter as a useful base class, and it provides almost
no functionality.  So it mainly described a pattern, but variations in
different filters make compilers unhappy.  So they now are a set of classes
with similar names and concepts (a sequence of value updates, and a way to find
stuff out about those values).


------------------------------------------------------------------------
r1205 | ckillian | 2012-05-16 14:38:32 -0400 (Wed, 16 May 2012) | 10 lines