Making ContextMapping class part of a member of service class.
Previously, ContextMapping is declared as a global static object. However, it limits the use of service classes. For example, it will not support service composition because it does not know which context mapping relationship is associated with which services. If multiple instances of the same type of service class is created, they would use the same mapping, which is not desirable.