Attempt to fix race condtion.
Race condition occurs because in a context possessed by an event, the runtime keeps track the set of child contexts below this context that is accessed so far. Originally this mechanism was developed to support dynamically created contexts. But because the set of child contexts is stored as a member variable of the context object, if multiple events executing on the same context, it's possible latter events add new child contexts to the child context set, affecting the assuming behavior of the earlier events.

This fix attempts to address this bug by: at the beginning of event creation, makes a copy of initial set of child contexts in a thread-specific space. Whenever a new child context is accessed, it not only adds a new entry into the context object, but also adds one into thread-specific space. At the end of the event when the event sends commit message to child contexts, it uses the local copy of child context set stored in thread-specific area.

Not tested yet. Will test to make sure it works.