Attempt to solve the problem of context fan-out.
Conceptually, in a hierarchical context structure, each context has a number of parent contexts, called fan-in, and a number of child contexts, called fan-out.

Assuming a tree-like context hierarchy, that is, fan-in, of all contexts except global context, is 1. So we will worry about fan-in later. Let's talk about fan-out. Currently we are assuming a static mapping of contexts to physical nodes, that is, the set of contexts is pre-determined. But eventually we will extend it so that contexts can be created dynamically.

Dynamical creation of contexts is non-trivial. 1. Suppose a context c is created by event x, then the now_serving and now_committing number will start from x, instead of 1. 2. The run-time library have to assign a physical node to host the context c and to inform the parent context of c to update fan-out child context. If DAG-like structure is assumed, that is, fan-out > 1, I can not think of any way to do the pointer 2. Even tree-like is assumed, there is still problem. Suppose the structure

   g
  / \
 x  y

If an event possesses the context x and y, and is currently in context x (meaning the event is currently executing in a function with target context x) and the event accesses a new child context z of y. At context x the event will have to create a new mapping, broadcast the mapping update to all physical nodes, update y's fan-out and create z context. This adds a lot of complexitities.

To ease the solution, the current design assumes when a context is accessed the first time, it has to be specified as the target context of the event. The head node inspects the target context ID, if it finds the ID appears the first time, it creates the context mapping and creates a new event (before the real event) to update the fan-out of the parent of the target.