Added ThreadStructure::myEvent() and added a headContext.
ThreadStructure::myEvent() is the event id of the highlevel event. It differentiates from ThreadStructure::myTicket() which is the local event ticket number.

headContext is added so that head node operation can run in parallel with other nodes if the physical node that hosts head also hosts other context (e.g. when all contexts are consolidated into single physical node)