1. Use async queue for local messages. 2. Flatten upcall messages
(1) This revision uses async dispatch threads to deliver messages that are sent locally.
Currently, not all local messages are sent via async dispatch thread, but I will try to make as much as possible.

(2) When there is a context'ed upcall event handler, the message is actually wrapped into another message.
Previously, the message is directly wrapped, but now the message is "flattened" into the wrapper message.