Added a 'vnode_MaceKey', a subclass of MaceKey which is used to designate a virtual node.
This class contains a uint32_t nodeid, which is assigned when job scheduler assigns a new job.

Internally, there should be a lookup map that converts virtual node id into actual MaceAddr address. (not done yet) Job scheduler is responsible for updating the mapping when the new job is created and when the job is killed or terminated.