For the message in the context'ed upcall deliver handler, the corresponding downcall_route() is modified so that the message is redirected to local head.
At the head, it creates a new event, and send to remote node.

For each such message, there are two more auxiliary message being created. One is created to be sent to the local head, and the other is created so that
when the message is received at the head, the head automatically creates a receive message event and send to the corresponding context.