Adding test cases.
Added a test app in applications/cjtest. In this directory, an executable 'driver' is used to emulate the input from the job scheduler. Alternatively, cjtest can run in standalone mode.

In addition, in ServiceImpl.pm, I added a helper method called 'asyncHead()' which contains code that runs at head node and are the same for every async/new events.