Changed compiler in order to make FileSync service to compile.
However, FileSync service was also edited. Some of the change was because the current compiler is more strict.
(e.g., upcall deliver( src, dest, const Message& msg ){ } --> not support syntax now )

Some are because returning references:

const MaceKey& getMaster( ){ } --> this does not compile and I do not think it should be supported. Returning a reference only makes sence when the context of the routine is at the same process with the caller function/transition.

const MaceKey getMaster() const {}  --> this is a bug in the compiler that renders this case uncompilable. I will address this issue in the next revision.