ScopedContextRPC does not use context id to identify the thread to be waken up.
(1)Instead of context id, ScopedContextRPC uses the event id as the identifier.
This change was made because when making a downcall/upcall, there's no such thing as "source context" unlike the routines.

(2) snapshot handler now takes an additional parameter, the event id.

(3) maceInit() now send message to head node just once. Suppose the maceInit() from the upper service calls the maceInit() of the lower service,
the runtime determines that the event id is already created and therefore will not message the head node to attempt to create a new event id.