Added teset case 3 which tests making a downcall into lower service.
There's a bug which would require a bigger change to the runtime to fix it.

The bug exists because when making a downcall/upcall/routine call, the runtime needs the "current" context id . But in the case of downcall/upcall transition, there's no such thing as the "current" context, because the transition originates from different services.

There are several possible fixes to this bug, and I'm assessing which one of them is better. I'm confident that before this week's meeting, I'll be able to resolve most of the problems in the service layering implementation.