Added TestCase 4 service to test upcall transition into upper service and into application.
The transition into application has to be either deferred or blocked, depending on the return type.
To do this, the registerHandler() helper function now takes an additional parameter of type bool. If this parameter is true (this is the default value),
the lower service will recognize the handler registered is actually in the application. The compiler has been modified so that any services will register
as the handler of the lower service with parameter value false to indicate the handler is actually a fullcontext service and therefore no deferring should
take place.

TODO: the actually deferral operation is not done yet. There's a place holder labeled "TODO", and I will start to work on that in the next revision.