Attempt to fix (1) race condition after context is migrated (2) segfault when program terminates
(1) When migration takes place, head node sends an AllocateContextObject before any request is sent. This ensures no later events can enter the new node and create context object with the wrong now_serving number.

Set default number of async dispatch threads to 8. This helps reduce deadlocks & program freezing.

(2)
There was a bug where the main thread is ready to terminate, the other threads aren't aware of it and still atempt to execute. The result is a segfault because the main process already releases memory but other threads still tries to access it. The dirty solution now is to send SIGTERM to the process itself.

However, a more cleaner solution would be to tell the head node to send a "MACE_EXIT" event to all physical nodes, which prevent head node to issue any more events, and let it wait for remaining events to finish and gracefully terminates.