Bug fix: making the runtime more robust (see details below)
modifed migrationtest to add more complicated testing pattern.
The user supplied default context mapping is stored in defaultMapping. The behavior is changed now.

Removed the use of headMutex temporarily. The existence of headMutex was to prevent block the system while head node is doing something. However, it seems sometimes causes deadlock.
I will restore this part once the runtime becomes stable.

When a new node joins the logical node the first time, (by  context migration ) the current context mapping is also transfered.

TODO: when a new context is created, and is the first context on that physical node, transfer current context mapping to that node.