A partial fix for a bug (described below)
This bug only occurs when several services are composed together. When a new context is created in one service, a NEWCONTEXTEVENT is created and propogated in that service context hierarchy. However, that only creates mapping snapshot in that service. Other service when attempting to access that snapshot will not find the corresponding snapshot and fail the system.