Bug fix for composed service maceExit()
The bug was that when the first maceExit()(in the topmost service) is called, when that maceExit transition returns it not only commits the contexts in that service, but all services in the application.

When the non-topmost services enters maceExit() subsequently, it will find the end event already commit at the contexts and then abort.

This bug doesn't affect the execution of the application, until the application exits.