An attempt to fix a bug discovered today:
When an event transition into a service for the first time, it may not enter the global context.
Instead, it may enter a lower context. It created problem with the runtime because runtime can not correctly reason
what are the set of contexts that the event has entered and can't downgrade contexts correctly.

This changeset serves as an attempted fix. Need verification.