Bug fix
For descendent context object of a context object, a pointer points from the context object to the descendent context object. However, the context initialization did not initialize the pointer so the pointer address is invalid (it's not NULL as I supposed it to be), which causes the problem when traversing the context hierarchy using the context canonical name.

It is fixed now by initializing the pointer to NULL.