Attempt to fix thread consumption problem.
The old design has the problem that when an event acquires the lock to a context, it also has to inform other contexts that it does not enter those contexts.

When doing so, the message handler has to wait for the previous events to downgrade before it can downgrade. Therefore, a huge number of threads is used and wait
simply to downgrade itself. This is not efficient and consumes much more threads than it would otherwise need.

The new design simply add two ticket sets that when the completed event is ready to signal the next event, it checks to see if that set is empty, if not, it increment
now_serving or now_committing counters. In addition the event that enters in NONE_MODE does not block waiting. It simply insert the ticket into the queues.