Bug fix.
This changeset fixes a bug that existed in the last commit, that in some cases a waiting thread may not be signaled when it's supposed to be signaled.

Plus, a new feature that reduces the number of threads being used per event, it can now handle 700 contexts on a machine if MAX_ASYNC_THREADS is set to 3000.

This should be more than enough for our purpose. Though I can imagine a better fix can increase the max number of contexts.

I tested this code with single node Tag service. It seems there's a bug with multi-node configuration...