Ok! Hopefull this will work.
With this new code, the runtime can limit the number of events that can proceed.

Specifically, the parameter "MAX_UNCOMMITTED_EVENTS" regulates the events that are uncommitted.

By default, this value is MAX_ASYNC_THREADS/10.


Use it with caution.... I ran it several times without failure but I can't guarantee the program
logic is 100% correct. Concurrency is a monster.