Runtime change:
When an event finishes, it should first downgrade all contexts before report at head node.

This reduces the overhead at the head, as well as enhancing parallelism because the head commit
code uses AgentLock to protect shared variables, and this new code does not require that.

Performance further improved to 35 seconds for 60 clients( down from 150 seconds)