Make ThreadPool slightly more efficient.
Previously, each thread in ThreadPool records its state (sleep/busy) in sleeping[] array. When testing if all the available threads are busy, the runtime had to travese the entire array to determine. This is stupid when the number of threads are large : for example, more than a thousand threads (which is typical).

So instead, I just maintain a sleeping counter, determining the business of the threads is simply a O(1) operation rather than O(n).

Empirically, there is a small performance improvement with this new approach.