Resolving strict-aliasing problem.
When compiling with -O2 flag (releasedebinfo)
g++ complained:
/scratch/chuangw/mace-fullcontext-onemsg/lib/MaceKey.h:1228: error: dereferencing type-punned pointer will break strict-aliasing rules

It seems in the recent g++ version, strict-aliasing is enforced by default with -O2 flag.

I resolved the compilation error by assigning __attribute__((__may_alias__)) to the variable, but I don't know if compilation would do
something funny. Anyway, that part of the code is not used currently with fullcontext.