Make perl compiler much much faster.
I've been annoyed that the perl compiler has been slow since some time before. It is not until today that I realized the "given/when" statement added to replace if/else statement in revision 298 was the trouble.

Specifically, perl has a new switch/case-like statement called 'given'/'when'. There are two given/when implementation in perl. One is the older, (much) slower Switch.pm implementation, and a newer, intrinsic implementation that appeared since perl 5.10. Unfortinately, when I first learned to use given/when, I used the older one. It is very slow because it was intended to be used by the old perl interpreter which did not understand the syntax. The Switch.pm module is based on 'source filtering', that is, it parses the perl code and translate it. Because our perl compiler code is very huge, thousands of lines of code, the translation step is extremely slow. It took around 10 seconds for it to load every module before, and now after the update, it takes just 2 seconds to do it.