A temporary hack to resolve an upcall rid problem.
Some legacy mace services uses the "authortativeJoinHandler" approach to determine the registration id of the upcall destination.
However, this approach is problematic in fullcontext mace, because the approach stores the registration id in a state variable.
In full context model, state variables are in global context, and it's not feasible to acquire the global context snapshot simply
to send an upcall.

This temporary hack looks to see if the remapped transition default registration id is one of the state variables. If so, it
set a pointer to global context object.

Therefore, this hack will still not be applicable to fullcontext services. It is provided for backward compatibility.