Moved several context message handler to ContextService.
Specifically, messages AllocateContextObject, ContextMigrationRequest and TransferContext

Moving the C++ code from inside the perl compiler into lib/ will help a lot in the future.
It will not require to understand the structure of the compiler in order to modify the message handler,
and it also reduces the size of compiler.

--HG--
branch : chuangw/multicompiler