Service unit testing!
It is the initial step towards unit-testing a service. The idea is to create a minimalist service, and test it to make sure the basic APIs do work.

In the future, it can also develop into a full set of ContextLattice API in C++, where it is (fairly) easy to write full context Mace in C++ only.

--HG--
branch : chuangw/unit_test