Added a paramter to disable Nagle's algorithm.
I found there was a problem when sending small packets, packets are buffered for 40ms before it arrives at the destination, even on the high speed LAN.

After diagnoses, I concluded it was not the application's fault, and it's the Nagle's algorithm to blame.

SO I made the tcp service can configure to disable Nagle's algorithm.

Example:

# disable Nagle's algorithm
SET_TCP_NODELAY = 1

This parameter is 0 by default. (on Linux, Nagle's algorithm is enabled by default)

--HG--
branch : chuangw/unit_test