Big performance improvement!
With this changeset, I was able to raise the number of processed event from 36,000 per second to 62,000 per second.

Basically, the change is:

(1) AsyncDispatch thread no longer get a new ticket.
A new ticket is given only when an event is created. Therefore, it no long requires the hack to mark the ticket as not being used.
Subsequently, every async dispatch thread does not need to lock the AgentLock::__agent_ticketbooth global mutex any more.

Performance improved by 10%, from 36,000 events/s to 40,000 events/s.

But the real big improvement comes from the following:
(2) Previously, to create a new event, the parent event signals a thread from ThreadPool and give the event request to that thread.
In that scheduled thread, it simply put the request into event request queue (there is a thread dedicated to creating events), and leave.

It seems context switch is the problem. I found since ThreadPool is the bottleneck, the parent event can can directly put the event
request into the event request queue, and therefore reduce a thread call. After this, the number of events bumped from 40,000 events
to 62,000 events per second.

This change applies to local events only. I will work on the transport service so that it can also take advantage of this improvement.

Note that further improvement is possible -- Current implementation when events entering a context will need to wait until the previous
events releases the lock, and the thread executing the event will block. I imagine if I use the SEDA-like model for executing events in context,
then threads will no longer need to wait. There will be lesser context switch penalty and better performance.